Upgrade to tclap 1.2.5
