+++ cpp/external/tclap-1.2.5/README
@@ -6,12 +6,12 @@ of help output.)
 TCLAP - Templatized Command Line Argument Parser
 
 This is a simple C++ library that facilitates parsing command line
-arguments in a type independent manner.  It doesn't conform exactly
-to either the GNU or POSIX standards, although it is close.  See
+arguments in a type independent manner. It doesn't conform exactly
+to either the GNU or POSIX standards, although it is close. See
 docs/manual.html for descriptions of how things work or look at the
 simple examples in the examples dir.
 
-To find out what the latest changes are read the NEWS file in this directory.
+To find out what the latest changes are read the NEWS file in this
+directory.
 
-
-Any and all feedback is welcome to:  Mike Smoot <mes@aescon.com>
+Any and all feedback is welcome at https://sf.net/p/tclap/discussion/
