+++ cpp/external/tclap-1.2.5/include/tclap/CmdLineOutput.h
@@ -1,3 +1,5 @@
+// -*- Mode: c++; c-basic-offset: 4; tab-width: 4; -*-
+
 
 
 /****************************************************************************** 
@@ -5,6 +7,7 @@
  *  file:  CmdLineOutput.h
  * 
  *  Copyright (c) 2004, Michael E. Smoot
+ *  Copyright (c) 2017, Google LLC
  *  All rights reserved.
  * 
  *  See the file COPYING in the top directory of this distribution for
