+++ cpp/external/tclap-1.2.5/include/tclap/DocBookOutput.h
@@ -5,6 +5,7 @@
  *  file:  DocBookOutput.h
  * 
  *  Copyright (c) 2004, Michael E. Smoot
+ *  Copyright (c) 2017, Google LLC
  *  All rights reserved.
  * 
  *  See the file COPYING in the top directory of this distribution for
@@ -100,7 +101,7 @@ inline void DocBookOutput::usage(CmdLineInterface& _cmd )
 	std::string xversion = _cmd.getVersion();
 	theDelimiter = _cmd.getDelimiter();
 	XorHandler xorHandler = _cmd.getXorHandler();
-	std::vector< std::vector<Arg*> > xorList = xorHandler.getXorList();
+	const std::vector< std::vector<Arg*> > xorList = xorHandler.getXorList();
 	basename(progName);
 
 	std::cout << "<?xml version='1.0'?>" << std::endl;
@@ -134,7 +135,7 @@ inline void DocBookOutput::usage(CmdLineInterface& _cmd )
 
 		std::cout << "</group>" << std::endl;
 	}
-	
+
 	// rest of args
 	for (ArgListIterator it = argList.begin(); it != argList.end(); it++)
 		if ( !xorHandler.contains( (*it) ) )
@@ -242,6 +243,7 @@ inline void DocBookOutput::printShortArg(Arg* a)
 		removeChar(arg,']');
 		removeChar(arg,'<');
 		removeChar(arg,'>');
+		removeChar(arg,'.');
 		arg.erase(0, arg.find_last_of(theDelimiter) + 1);
 		std::cout << theDelimiter;
 		std::cout << "<replaceable>" << arg << "</replaceable>";
@@ -280,6 +282,7 @@ inline void DocBookOutput::printLongArg(Arg* a)
 		removeChar(arg,']');
 		removeChar(arg,'<');
 		removeChar(arg,'>');
+		removeChar(arg,'.');
 		arg.erase(0, arg.find_last_of(theDelimiter) + 1);
 		std::cout << theDelimiter;
 		std::cout << "<replaceable>" << arg << "</replaceable>";
