+++ cpp/external/tclap-1.2.5/include/tclap/HelpVisitor.h
@@ -1,3 +1,5 @@
+// -*- Mode: c++; c-basic-offset: 4; tab-width: 4; -*-
+
 
 /****************************************************************************** 
  *
