+++ cpp/external/tclap-1.2.5/include/tclap/StdOutput.h
@@ -5,6 +5,7 @@
  *  file:  StdOutput.h
  * 
  *  Copyright (c) 2004, Michael E. Smoot
+ *  Copyright (c) 2017, Google LLC
  *  All rights reserved.
  * 
  *  See the file COPYING in the top directory of this distribution for
