+++ cpp/external/tclap-1.2.5/include/tclap/XorHandler.h
@@ -1,3 +1,5 @@
+// -*- Mode: c++; c-basic-offset: 4; tab-width: 4; -*-
+
 
 /****************************************************************************** 
  * 
@@ -55,8 +57,8 @@ class XorHandler
 		 * Add a list of Arg*'s that will be xor'd together.
 		 * \param ors - list of Arg* that will be xor'd.
 		 */
-		void add( std::vector<Arg*>& ors );
-			
+		void add( const std::vector<Arg*>& ors );
+
 		/**
 		 * Checks whether the specified Arg is in one of the xor lists and
 		 * if it does match one, returns the size of the xor list that the
@@ -84,7 +86,7 @@ class XorHandler
 		 */
 		bool contains( const Arg* a );
 
-		std::vector< std::vector<Arg*> >& getXorList(); 
+		const std::vector< std::vector<Arg*> >& getXorList() const;
 
 };
 
@@ -92,8 +94,8 @@ class XorHandler
 //////////////////////////////////////////////////////////////////////
 //BEGIN XOR.cpp
 //////////////////////////////////////////////////////////////////////
-inline void XorHandler::add( std::vector<Arg*>& ors )
-{ 
+inline void XorHandler::add( const std::vector<Arg*>& ors )
+{
 	_orList.push_back( ors );
 }
 
@@ -150,7 +152,7 @@ inline bool XorHandler::contains( const Arg* a )
 	return false;
 }
 
-inline std::vector< std::vector<Arg*> >& XorHandler::getXorList() 
+inline const std::vector< std::vector<Arg*> >& XorHandler::getXorList() const
 {
 	return _orList;
 }
