+++ cpp/external/tclap-1.2.5/include/tclap/ZshCompletionOutput.h
@@ -237,8 +237,9 @@ inline void ZshCompletionOutput::printOption(Arg* a, std::string mutex)
 	if ( a->isValueRequired() )
 	{
 		std::string arg = a->shortID();
-        // Example arg: "[-A <integer>] ... "
-        size_t pos = arg.rfind(" ... ");
+        // Example arg: "[-A <integer>] ..."
+        size_t pos = arg.rfind(" ...");
+
         if (pos != std::string::npos) {
             arg.erase(pos);
         }
