+++ cpp/tests/results/cmd/match_help.stdout
@@ -6,7 +6,7 @@ DESCRIPTION:
 
 USAGE: 
 
-   match  -config <FILE> ...  [-log-file <FILE>] [-sgf-output-dir <DIR>]
+   match  -config <FILE> ... [-log-file <FILE>] [-sgf-output-dir <DIR>]
           [...other flags...]
